Blue charge pump perfect! Pixel ISR perfect!
Finally sat down and got that all correct and the code is lovely. Trick
was (1) do each color separately, and (2) have a rest phase between
pixels for the charge pump.  This lowers peak current draw too since
only one LED is ever on at any given moment. Costs is lower refresh
rate, but looks fine to me. We do need to bump up the gamma values since
the duty cycle is now 1/4th what it was, but still looks good.