Ugly workaround to keep Arduino IDE from getting errors compiling the Serial class
Apparently IDE compiles evrything in the directory, even if it is not #included.  WTF. This means that the Serial was getting compiled even when you didn't include it at all, and that made errrors since there was no hard definitions in the legacy variant. Arg. Solution is to define fake hardware in the variant.  So ugly. Hopefuly that variant will go away someday, but this has to come up again someday.