Implements double buffered display in core and blinklib.
In the core, `pixel_setRGB()` has been replaced with `pixel_bufferedSetPixel()`. All pixel updates are now double buffered and the changes are not displayed until `pixel_displayBufferedPixels()` is called. The update is also synchonized the the to next display frame so every call to `pixel_displayBufferedPixels()` will be displayed for at least one full frame, and partial frames will never happen.

In blinklib, the impact is that all pixel changes made durring loop() are batched and the cumulative result is applied atomically each time loop() returns. This breaks code that used `delay()` between updates in the same loop() iteration -> which is desired since this is not best practice. All the now broken examples need to be fixed.