Rewrite of IR recieve code to avoid `case` statement
The `case` was clean, but it was getting complied awefully and blowing the stack.
Cleaned up logic and made much more efficient. Reduced the state count.
Also got rid of the `IDLE` requirement at the begining of a byte because a spontainously trigger could happen right before a byte and thus make you miss it. The `IDLE` at the end shoudl catch all mallformed bytes.