Use GCC depreicated attribute to warn `setFaceColor()` calls
The previous code used a `#warning` which would actually fire every time the .cpp file was compiled even if the dpreicated call was not used. It also would not fire on a call to the depricated function in another file.

Unfortunately there is no way to get a hint with this method, but at least you (1) don't get the warning if you don't call the depricated function, and (2) do get the warning if you do call the depricated function.  :)