Add missing line from the blinklib error correction encoding lookup table!
Oh man what a bug! This table was generated using the code above it. I used an online C compiler to print it. When coping/pasting the output from the website into the code, it lost the line that happened to land on the page break!!!!
Thank god the game BossFight just happened to use the exact missing value of 34 in a plainly manifest way or we would have never ever ever found this!!!!!