Download with retries works good. Not always the best download structure, but usually good.
Maybe add a FINISHED message to short circuit the active timeout which now can take a couple seconds?

There is also a case where a root sometimes times out when next to a child with multipule children, need to track that down and try to find a way to avoid it.