Faster timeout at end, but still takes 1 sec per hop to proigate back.
TODO only accept new children for limited window at begining of seed, and then we can keep track of which children are done compared to the list of known ones to know when we are done without needing to wait. Use a bitfield for finished flags so we can quickly XOR them with active bitfield to see if we are done.