Shuffle functions around to free up as much subbls space as possible for real games
We are now using everything above 0x3000. That leaves 0x1800 (6,144 dec) bytes for each the built-in and active games.  The built-in game starts at 0x1800.

Soon we will pull all the pixel and IR stuff into the bootloader so games will not have to dupe that code, giving them much more space.