Fixed makeColorHSB to use makeColorRGB to get correct 5 bit color values.
makeColorHSB() was using 0-255 for math and then directly putting this into a Color() which only can hold 5 bit values. Now uses a call to makeColorRGB() to convert the 8-bit values down to 5-bit values in the Color type returned.

How could this have been so wrong an no one noticed?