Now with optimal* program download linear path finding
I think this represents to optimal single pass path finding alogothim possible. It does remarably well. The only time it fails is in cases where it could not have known there was a shorter path on a more-clockwise branch.

Note that this release requires a manual button press on each blink after game download completes to actually start the newly downloaded image. You know the download is done when all blinks turn green and stop scanning.

Note that this release depends on timeouts to propigate completed terminal nodes back to the root, and this takes about 1 second per level of tree height. We can easily do an active backpropigation, but will need to decide if it is worth the code space.

(*I think :) )