Show a very fast blue spin dirring the 1 second release-to-seed window
It would be hard for user code to replicate this look becuase it is so fast. Helps user know when to release to enter seed mode. Just having blue is not good enough becuase the game could be showing blue. Also added instant feedback when you release becuase the checksum lag didn't feed responsive. .