Dither sleep/wake animations with packet sending.
We now do ghetto copperative multitasking to send the warm sleep/wake packets while also showing the animations steps so things feel more responsive.

It is all hardcoded and just happenes to work out to take aproximately the right ammount of time to send the 30 packets so the animation looks reasonable.

Luckily all this uglyness is nno-user facing and burried in blinklib so we are not locked in if we ever want to change it.

Before:
Sketch uses 1988 bytes (35%) of program storage space.

After:
Sketch uses 1904 bytes (33%) of program storage space.