Only check for low battery voltage at power up and when waking
Check for 2.1V on initial power up & on wake from deep sleep. It does this to make sure there there is sufficient voltage to switch to 8Mhz operation.

Except for no green 2.1V seems to work, at least in my blinks. Will take more testing to see if this is too low.

This version also cleans up all the async ADC code to hopefully make room a flag to see if you were root on the last download seed.

Finally, clears the pixels on wake to get rid of that annoyon simetimes-stale flash. Now the forground game has a chance to update the display after wake before we show anything.