Boards.txt and platform.txt updated for new blink328 board type
Oh man that took so many hours. The IDE was somehow using lines from the JSON installed platform.txt and mixing them in with the manually installed one. What a mess. Solution was to completely remove the JSON one and restart and then it got very confused and forgot everything... but then used everything from the correct platform.txt. Yuck.

Seems to work and program right, we still need to fix the BlinkBIOS code to know where the active and built-in games are.