Switch 328->328PB. Force interrupt vector table on 328 to be at the same address as the one on 168
The BIOS makes a duplicate vector table at 0x3800 on the 328 so that there is always a table here om both parts (on the 168 the table at 0x3800 is the real one inside the bootloader).

This change makes games compiled for the 328 target use this fake table so that if there ever get run on a 168 (because of IR download) then they will still find a table at this address.