Give the linker a little more room that it should not need
It seems that the linker sometimes overflows .text prematurely. We give it a little extra room here so that it does not bail. This will make a few games that are right on the edge be able to compile and they do end up being the right size in the end. For those that actually end up being too large to fix, we will catch those with the final size check in the Arduino IDE build process (uses `upload.maximum_size` in `boards.txt`)