Fix `platform.txt` to make exported binaries include the `.data` section
You really need the `.data` section or any program that has initialized data in it (nost!) will not work because the data will not be inside the binary image.
Note that you still do need to *remove* the `.bss` section since that really should not be in there (it is not even the correct kind of memory since bss in in SRAM), so change the command line to *exclude* bss rather than *include* data since if we only include `.text` (yep, the command is "include *only*" so you can not explicitly include multiple sections.   the we do not get the `.data` section.  So we just got lucky that there is only one section we need to exclude.